requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Faker==12.0.1
idna==3.3
importlib-metadata==6.0.0
Markdown==3.4.1
mixer==7.2.2
pygame==2.1.2
python-dateutil==2.8.2
pytz==2022.7
requests==2.27.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.9
wcwidth==0.2.5
zipp==3.11.0
